+++ src/columns/TextInput.js
@@ -23,6 +23,7 @@ class TextInput extends Component {
 
   onChangeHandler = (event) => {
     // TODO: Debounce events? lodash.debounce(...)
+    // console.log('TextInput value ', event.target.value, 'id', this.props.id);
     this.props.sendValue(event.target.value, this.props.id)
   }
