+++ src/components/RegistrationCollection.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import RegistrationRow from '../registration-row/RegistrationRow';
+import RegistrationRow from './RegistrationRow';
 
 const defaultCollectionStyle = {
   border: 'solid thin black',
